+++ packages/svelte/src/compiler/phases/2-analyze/visitors/ClassBody.js
@@ -7,21 +7,24 @@ import { get_rune } from '../../scope.js';
  * @param {Context} context
  */
 export function ClassBody(node, context) {
-	/** @type {string[]} */
-	const private_derived_state = [];
+	/** @type {{name: string, private: boolean}[]} */
+	const derived_state = [];
 
 	for (const definition of node.body) {
 		if (
 			definition.type === 'PropertyDefinition' &&
-			definition.key.type === 'PrivateIdentifier' &&
+			(definition.key.type === 'PrivateIdentifier' || definition.key.type === 'Identifier') &&
 			definition.value?.type === 'CallExpression'
 		) {
 			const rune = get_rune(definition.value, context.state.scope);
 			if (rune === '$derived' || rune === '$derived.by') {
-				private_derived_state.push(definition.key.name);
+				derived_state.push({
+					name: definition.key.name,
+					private: definition.key.type === 'PrivateIdentifier'
+				});
 			}
 		}
 	}
 
-	context.next({ ...context.state, private_derived_state });
+	context.next({ ...context.state, derived_state });
 }